Format code

+++ contracts/interfaces/IOracle.sol
@@ -2,6 +2,10 @@ pragma solidity ^0.6.0;
 
 interface IOracle {
     function update() external;
-    function consult(address token, uint amountIn) external view returns (uint amountOut);
+
+    function consult(address token, uint256 amountIn)
+        external
+        view
+        returns (uint256 amountOut);
     // function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestamp);
 }
