+++ contracts/lib/Babylonian.sol
@@ -1,10 +1,10 @@
 pragma solidity ^0.6.0;
 
 library Babylonian {
-    function sqrt(uint y) internal pure returns (uint z) {
+    function sqrt(uint256 y) internal pure returns (uint256 z) {
         if (y > 3) {
             z = y;
-            uint x = y / 2 + 1;
+            uint256 x = y / 2 + 1;
             while (x < z) {
                 z = x;
                 x = (y / x + x) / 2;
